Admin messages view & upd bot parse only public chats

+++ db/migrate/20230313171051_create_messages.rb
@@ -2,8 +2,8 @@ class CreateMessages < ActiveRecord::Migration[7.0]
   def change
     create_table :messages do |t|
       t.string :message_id
-      t.integer :chat_id
-      t.integer :sender_id
+      t.bigint :chat_id
+      t.bigint :sender_id
       t.datetime :datetime
       t.json :text
       t.text :original_message
